refactor: move filter.js to js package #16
